Add appendMeals to domIndex.js and make SRP
