+++ lib/domIndex.js
@@ -13,7 +13,7 @@ const remainingCaloriesClassType = (goal, total) => {
   }
 }
 
-export const remainingCalories = (tableName, total) => {
+const remainingCalories = (tableName, total) => {
   let goal = mealCalories[tableName]
   let type = remainingCaloriesClassType(goal, total)
   $(`#${tableName}`).append(`<tr>
@@ -22,7 +22,7 @@ export const remainingCalories = (tableName, total) => {
   </tr>`)
 }
 
-export const dayTotalCalories = (total) => {
+const dayTotalCalories = (total) => {
   let type = remainingCaloriesClassType(2000, total)
   $(`#meal-totals`).html('')
   $(`#meal-totals`).append(`<tr>
@@ -39,9 +39,40 @@ export const dayTotalCalories = (total) => {
   </tr>`)
 }
 
-export const mealTotalCalories = (tableName, total) => {
+const mealTotalCalories = (tableName, total) => {
   $(`#${tableName}`).append(`<tr>
   <td>Total Calories</td>
   <td>${total}</td>
   </tr>`)
+}
+
+const appendMealsTableHead = () => {
+  $('#breakfast').html('<caption>Breakfast</caption><tr><th>Name</th><th>Calories</th></tr>')
+  $('#snack').html('<caption>Snacks</caption><tr><th>Name</th><th>Calories</th></tr>')
+  $('#lunch').html('<caption>Lunch</caption><tr><th>Name</th><th>Calories</th></tr>')
+  $('#dinner').html('<caption>Dinner</caption><tr><th>Name</th><th>Calories</th></tr>')
+
+}
+
+export const appendMeals = (meals) => {
+  let dayTotalCal = 0
+  appendMealsTableHead()
+  meals.forEach((meal) => {
+    let name = meal.name.toLowerCase()
+    let mealTotalCal = 0
+    meal["foods"].forEach((food) => {
+      $(`#${name}`).append(`<tr class="${food.id}">
+      <td class="food-${food.id}">${food.name}</td>
+      <td class="calories" class="cal-${food.id}">${food.calories}</td>
+      <td><button class="delete-mealFood" aria-label="Delete">
+      Delete
+      </button></td>
+      </tr>`)
+      mealTotalCal += food.calories
+    })
+    dayTotalCal += mealTotalCal
+    mealTotalCalories(name, mealTotalCal)
+    remainingCalories(name, mealTotalCal)
+  })
+  dayTotalCalories(dayTotalCal)
 }
