+++ lib/index.js
@@ -1,4 +1,4 @@
-import { remainingCalories, dayTotalCalories, mealTotalCalories } from './domIndex'
+import { appendMeals } from './domIndex'
 
 const validateForm = () => {
   removeValidations()
@@ -194,39 +194,6 @@ const getAllMeals = () => {
     .then((meals) => appendMeals(meals))
     .catch((error) => console.error({error}))
 }
-
-const appendMeals = (meals) => {
-  let dayTotalCal = 0
-  $('#breakfast').html('<caption>Breakfast</caption><tr><th>Name</th><th>Calories</th></tr>')
-  $('#snack').html('<caption>Snacks</caption><tr><th>Name</th><th>Calories</th></tr>')
-  $('#lunch').html('<caption>Lunch</caption><tr><th>Name</th><th>Calories</th></tr>')
-  $('#dinner').html('<caption>Dinner</caption><tr><th>Name</th><th>Calories</th></tr>')
-  meals.forEach((meal) => {
-    let name = meal.name.toLowerCase()
-    let mealTotalCal = 0
-    meal["foods"].forEach((food) => {
-      $(`#${name}`).append(`<tr class="${food.id}">
-      <td class="food-${food.id}">${food.name}</td>
-      <td class="calories" class="cal-${food.id}">${food.calories}</td>
-      <td><button class="delete-mealFood" aria-label="Delete">
-      Delete
-      </button></td>
-      </tr>`)
-      mealTotalCal += food.calories
-    })
-    dayTotalCal += mealTotalCal
-    mealTotalCalories(name, mealTotalCal)
-    remainingCalories(name, mealTotalCal)
-  })
-  dayTotalCalories(dayTotalCal)
-}
-
-// const mealTotalCalories = (tableName, total) => {
-//   $(`#${tableName}`).append(`<tr>
-//   <td>Total Calories</td>
-//   <td>${total}</td>
-//   </tr>`)
-// }
 
 const createFoodInMeal = (event) => {
   event.preventDefault()
